+++ src/SciTEIO.cxx
@@ -359,7 +359,7 @@ static const char codingCookie[] = "coding";
 static UniMode CookieValue(const SString &s) {
 	int posCoding = s.search(codingCookie);
 	if (posCoding >= 0) {
-		posCoding += strlen(codingCookie);
+		posCoding += static_cast<int>(strlen(codingCookie));
 		if ((s[posCoding] == ':') || (s[posCoding] == '=')) {
 			posCoding++;
 			while ((posCoding < static_cast<int>(s.length())) &&
